Rename module from work withs db
