+++ app/orders/views.py
@@ -67,7 +67,7 @@ from .utils.files import (
     upload_image_to_answer,
 )
 from .utils.order_state import OrderStateActivate
-from .utils.clone_db_data import CloneOrderDB
+from .utils.db_data import CloneOrderDB
 
 
 @swagger_auto_schema(**swagger.OrderCreate.__dict__)
